project/requirements.txt
appdirs==1.4.4
attrs==20.1.0
black==19.10b0
click==7.1.2
flake8==3.8.3
mccabe==0.6.1
pathspec==0.8.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
regex==2020.7.14
toml==0.10.1
typed-ast==1.4.1
